+++ src/index.js
@@ -1,7 +1,4 @@
-import { example } from './example';
 import { display } from './js/interface';
-import { controller } from './js/controller';
 import './styles/main.scss';
 
 display.render();
-console.log(example('bro'))
