--- src/js/controller.js
@@ -1,10 +1,0 @@
-const controller = (() => {
-    const save = (key, elem) => {
-        localStorage.setItem(key ,JSON.stringify(elem));
-    }
-
-
-    return {save,}
-})();
-
-export {controller};
